feat: refactor MCP and auth handling by consolidating WebSocket and HTTP endpoints
